refactor: address AppStore review feedback
