+++ src/application/entities/mod.rs
@@ -9,5 +9,5 @@ pub use ordering::*;
 pub use services::*;
 pub use store::AppStore;
 #[cfg(test)]
-pub use store::StoreMocks;
+pub use store::MockAppStoreBuilder;
 pub use transaction::*;
